requirements.txt
flask
authlib
werkzeug
requests
python-dotenv
